pyproject.toml
[tool.poetry]
name = "kubectl-val"
version = "0.1.4"
description = "Run formal check of Kubernetes configurations using Poodle AI planner"
authors = ["CriticalHop Team <info@criticalhop.com>"]
license = "Apache-2.0"
readme = "README.md"

packages = [
    { include = "kalc" },
]

[tool.poetry.dependencies]
python = ">=3.7"
PyYAML = "^5.1.2"
yaspin = "^0.15.0"
Click = "^7.0"
logzero = "^1.5.0"
pyupdater = "^3.1"
poodle = ">=0.1.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
kubectl-val = "kalc.cli:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

